Move endpoint constants to the GLOBAL package
