+++ src/GLOBAL/endPoints.js
@@ -3,5 +3,5 @@ export const getNearByCity = (latitude, longitude, limit = 10, type = 'city', mi
   longitude = longitude > 0 ? `%2B${longitude}` : longitude;
   return `https://wft-geo-db.p.rapidapi.com/v1/geo/cities?location=${latitude}${longitude}&limit=${limit}&minPopulation=${minPopulation}&types=${type}`;
 }
-export const cityWeatherEndPoint = (cityLatitude, cityLongitude) =>
-  `https://api.openweathermap.org/data/2.5/weather?lat=${cityLatitude}&lon=${cityLongitude}&appid=6902c072c502ba82c29e7bb1ee6cb7ec`
+export const cityWeatherEndPoint = (cityLatitude, cityLongitude, unit) =>
+  `https://api.openweathermap.org/data/2.5/weather?lat=${cityLatitude}&lon=${cityLongitude}&appid=6902c072c502ba82c29e7bb1ee6cb7ec&units=${unit}`
